Add CLI options using yargs, update deps
Behave more like a typical *nix app and accept command-line options as well as positionals (using yargs). More options will be added as kbp functionality is expanded.

Also update dependencies - note that @types/node 17.x is not compatible with typescript 4.9.x, so for now, typescript level is set to ~4.8.4, but ideally both should be updated together.