Add hacky rotation support
In KBS, rotation origin is always the top-left of the line. .ass only has that origin if alignment is left. This initial release just uses the default origin instead of trying to calculate the correct one to emulate KBS. Once actual KBS behavior can be achieved, this will be promoted to the main feature branch.