Remove sort of lines by timestamp during conversion
I don't believe this sort provided any benefit other than perhaps "cleaner" looking files, and it removes the implicit layering you get by specifying a line later in the file. I don't believe there is a requirement in the .ass format to have the timestamps all in order. If there were, to achieve the same, the layer parameter would have to be set for each line explicitly based on the order in the .kbp file.