Remove unnecessary alignment info on lines
The style's alignment is now determined by the first usage rather than the last, and it's actually taken into account now. If the line's alignment matches the style's alignment, no \an tag is emitted. This does introduce an inconsistency if the position option is enabled but cdg is not and the first line of a style is not centered, but I don't think anyone is using that combination of options in the program, so it's more helpful to remove extraneous tags than fix that at this point.